WIP: временный коммит перед переключением ветки

+++ src/mission_publisher/src/mission_publisher.cpp
@@ -146,7 +146,7 @@ MissionPublisher::Mission MissionPublisher::parseMissionFromJson(const nlohmann:
 
     const auto& velocities = mission_json["velocities"];
     for (const auto& v : velocities) {
-        mission.velocities.push_back(v.get<float>());
+        mission.velocities.push_back(v.get<double>());
     }
     
     return mission;
